Remove generated files.
Specifically, this removes all generated files in the doc/src/user_guide files.
This includes the .pdf file, but there is still a version of the .pdf file of
the user guide in doc/ -- in other words, the repository did not just contain
one copy of the .pdf file, but two.