Used the mass conservation equation to increase numerical accuracy
In Diagnostic_TurbKE_Budget.F90 many of the routines computed the
divergence of the velocity field through the three diagonal components
of the rate of strain, div.u = e_rr + e_tt + e_pp. I replaced all
of these instances with a representation that uses the mass conservation
equation, div.u = -(1/rho) drho/dr v_r . This requires fewer derivatives
and should be more accurate.  Further, there was one place in which
the gradient of the divergence was needed. Use of mass conservation
eliminates the need for second derivatives and reduces the number of
code lines significantly.