Gentle_L_Conservation calculation
Calculates L_z and then modifies the real component of the n=0, l=1, m=0 mode of Z to remove it. Change is no more than 1 part in 1e6, and runtime with the fix is not distinctly longer than no treatment of L_z (3 minute difference after 3 hours, within normal variance).