add option to cause orderly shutdown by touching a file
This makes it possible to cause Rayleigh to exit and write a checkpoint
file by touching a file called "terminate" in the run directory. This is
convenient for restarting with new options without losing any progress.
With the right logic in the job script, even the current allocation in
the queue can be kept.

As checks for the existence of a file can be costly on some parallel
file systems, the default is to check only every 50 iterations. This can
be configured with `terminate_check_interval`.

The section "Running the code" in the documentation has been updated to
describe this feature.