remove line breaks in log file with ifort
ifort automatically inserts line breaks in long lines if they are
written with automatic formatting (*). This behavior can be overridden
by explicitly using character formatting '(A)'. I think this makes the
logs more consistent. It's better to let the editor/pager do the line
breaking for you depending on the size of your terminal.