enable machine-dependent SIMD optimizations also for GNU compilers
This uses a dialog similar to the one for the Intel compiler. I also
added an option to use automatic detection of the right instruction set,
provided the compiling node is the same as the compute node. The result
of auto-detect is displayed.

The option for producing several executables should work as well.
I chose the same selection as in the Intel compiler.