add configure option to use MKL provided by Debian/Ubuntu
Debian and Ubuntu had the intel-mkl package for several years now, which
provides a full MKL install as part of the normal package repository.
This flag makes it easy to use this MKL install. Just setting MKLROOT
doesn't work unfortunately because the paths can't be made to match for
libraries and includes.