An initial implementation of coupled bcs.  Untested but compiles.
Think this can actually be reimplemented in a considerably simpler way but I would like to add a test to this version first.

This may break current tests as it changes the default boundary condition logic (it is no longer possible to just set the bc type
based on what it is not because there are more than two types of bc now).